Update window.py
Create the window ui class, that contains an init function that will initalize the main start page window and start at the home page frame. Also created a create_frame class that destroys the previous frame and places another one where the old one was. This will be used to cycle through the different pages as the user clicks buttons and moves about the program.