Update deck_build_algo.py
Created a num_of_sources function that will determine, based on the mana value of a card and how many symbols are in its mana cost, how many mana sources of that type that deck will need to cast it reliably on time. 

Currently, the model only goes up to converted mana cost = 7, and num of symbols = 4, which should be good for the majority of cases, but there are a few exceptions, so over time I will gather more data and add it when I can.